requirements.txt
psycopg2-binary
PyJWT~=2.0.1
flask~=1.1.2
Werkzeug~=1.0.1
flask_restful